docs: Add programming-languages.rst
Most new contributors are probably going to modify existing code rather
than introducing all-new programs and scripts, and even when the latter
happen they'll hopefully get a feel for which programming languages are
considered acceptable for the project by looking at what's already in
the repo. Make this part less prominent by moving it to a separate page.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200406162010.332813-5-abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>