util: virdaemon: fix waiting for child processes
Unlike `waitpid`, `virProcessWait` only returns -1 (error) or 0
(success), so comparing that to `pid` will always be false and the
parent will report failure with:

error : main:851 : Failed to fork as daemon: No such file or directory

even though the grandchild process is succesfully running. Note that the
errno message is misleading: it was last set when trying to find a
restart state file.

Signed-off-by: Rafael Fonseca <r4f4rfs@gmail.com>
Message-Id: <20200407201434.1850252-1-r4f4rfs@gmail.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Reported-by: Marcin Krol <his-e-mail>