docs: add 'edit this page' link to footer of every page
To encourage contributors to make changes to the main website, add a
footer link to every page which links to the corresponding source file
in git. With gitlab, they are able to edit content directly in the web
browser and then submit a merge request. This gives a way to contribute
content that is arguably easier than our wiki which requires manual
account creation, while this will also benefit from maintainer review.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20200408152141.1182150-1-berrange@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>