util: virObjectLockable: convert virMutex to GMutex
Because virCond usage is connected to virMutex, this patch also converts
virCond to GCond.

An important note to make here: GMutex doesn't like when you try to
clear a lock that's still locked. So I had to hunt the few cases where
that happened and add an explicit virObjectUnlock call before unreffing
the object.

Signed-off-by: Rafael Fonseca <r4f4rfs@gmail.com>
Message-Id: <20200410135510.2148319-43-r4f4rfs@gmail.com>