conf: add xen specific feature: e820_host
e820_host is a Xen-specific option, only available for PV domains, that
provides the domain a virtual e820 memory map based on the host one. It
is required when using PCI passthrough and is generally considered safe
for any PV kernel. e820_host is silently ignored if set in HVM domain
configuration. See xl.cfg(5) man page in the Xen documentation for more
details.

Signed-off-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>

Message-Id: <4b30267da39a0c5bcc21c7d551c03e96433da617.1586831743.git-series.marmarek@invisiblethingslab.com>