qemu: fix domain start with corrupted save file
Commit 21ad56e932 introduced a regression where a VM with a corrupted
save image file would fail to start on the first attempt. This was
caused by returning a wrong return code as  'fd' was abused to also hold
the return code.

Since it's easy to miss this nuance introduce a 'ret' variable for the
return code and return it's value in the error section.

https://bugzilla.redhat.com/show_bug.cgi?id=1791522

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <0c8dc98cae4404d91966793d498815f8ba5ed741.1587567222.git.pkrempa@redhat.com>