mock: qemuDomainGetUnplugTimeout: Decrease timeout
We always queue the DEVICE_DELETED events before successful return from
the command so that tests are reliable. This means we can decrease the
unplug timeout as it's guaranteed to be executed in correct order.

According to my testing it shaves off ~450ms of test run:

real	0m0.721s

vs.

real	0m0.259s

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <e677932ab88121dae61c77f8d35a6821d57dc943.1587634711.git.pkrempa@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>