qemumonitortestutils: Enforce consumption of all items in test monitor
To prevent unexpected situations where a change in code would stop
looking at some of the tested commands go unnoticed add a mechanism to
force consumption of all test items.

Since there are a few tests which would be hard to fix add also a
mechanism to opt-out of the check.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <a2069faba2d1b41dda88d3f1ac02ea819dd0f427.1587655269.git.pkrempa@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>