libvirt-guests.sh: fix log output with new systemd
Newer systemd is too smart, it detects the PIDs of the gettext calls
and due to that a log lists libvirt-guests.sh with many different PIDs.

Furthermore it struggles to collect the output in time, so it can be
truncated and overall looks like:
   libvirt-guests.sh[37986]: Running guests on default URI:
   libvirt-guests.sh[37995]: Running guests on
   libvirt-guests.sh[37977]: R

Gather the gettext result into a local variable and printing
the value from libvirt-guests.sh itself fixes both issues.

Fixes: https://bugs.launchpad.net/ubuntu/+source/libvirt/+bug/1875708

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Message-Id: <20200430095957.2940867-1-christian.ehrhardt@canonical.com>