conf: Add qemu_job_wait_time option
The waiting time to acquire the lock times out, which leads to a segment fault.
In essence we should make improvements around locks, but as a workaround we
will change the timeout to allow the user to increase it.
This value was defined as 30 seconds, so use it as the default value.
The logs are as follows:

```
Timed out during operation: cannot acquire state change lock \
   (held by monitor=remoteDispatchDomainCreateWithFlags)
libvirtd.service: main process exited, code=killed,status=11/SEGV
```

Signed-off-by: MIKI Nobuhiro <nmiki@yahoo-corp.jp>
Message-Id: <1588316944-11240-1-git-send-email-nmiki@yahoo-corp.jp>