libxl: Use the name 'Xen' in driver tables
The libxl driver declares its name as 'Xen' through the public
virConnectGetType() API. In the virHypervisorDriver table the name is
set to 'xenlight'. To add more confusion, the name is set to 'LIBXL'
in the virStateDriver. For consistency, use the same name in the driver
tables as reported in the public virConnectGetType() API.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>
Message-Id: <20200504220919.23636-2-jfehlig@suse.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>