qemu: Don't use CPU from host capabilities as host-model on ARM
We never supported host-model CPUs on ARM and we don't want to support
them even once patches for direct detection of host CPU are merged. And
since using host CPU definition for host-model CPUs exist only for
backward compatibility, we should not use it for any host-model support
added in the future. Such enhancement should exclusively use the result
of query-cpu-model-expansion. Until proper host-model support is
implemented for ARM (if ever), we need to make sure the detected host
CPU is not accidentally used for host-model CPUs.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <f9bb07dd651a379c55cff5481db92b811bfcc9dc.1588752509.git.jdenemar@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>