qemu: block: add function to collect blockcommit params and verify them
Snapshot deletion is implemented as a three-phase process - first we
collect specifications of blockcommits used to perform the deletion and
run sanity checks to verify that the whole operation is reasonably safe,
then we launch the blockcommits and finally we wait for them to finish and
check the results.

This commit implements the first phase.  All information necessary to
launch and finish a blockcommit job per disk included in the snapshot is
collected in a blockcommit descriptor structure and checked to verify the
job makes sense and has a reasonable chance to succeed.  Broadly speaking,
the checks aim to ensure that the disks in the current running VM are the
same as they were when the snapshot was created.

Signed-off-by: Pavel Mores <pmores@redhat.com>
Message-Id: <20200506114226.2538196-7-pmores@redhat.com>