qemu: block: add external snapshot-delete top-level algorithm
This uses the helpers introduced in previous three commits and assembles
the top-level snapshot deletion algorithm in their terms.

Note that the third phase - waiting for blockcommits to finish - is
conceptually optional and a flag/parameter might be exposed to the user in
the future to skip the third phase.  This would make the whole operation
asynchronous and let the user deal with concluding the blockcommits
manually, using the generic blockjob tools.

Signed-off-by: Pavel Mores <pmores@redhat.com>
Message-Id: <20200506114226.2538196-10-pmores@redhat.com>